Fix #30 (#32)
This commit has fixed issue #30.
This commit added a special way for users who uses "nohup" command to run KookBC to stop the client instance.
See the log in the working directory, you will know.
(It is easy, write a file named KOOKBC_STOP to the working
directory, KookBC will found it and stop itself.)

------
Tested-by: ZX夏夜之风 <snwcreations@qq.com>