Event Builder, version 2 (#42)
WARNING: You should NOT revert this commit even if you need to create your own modded version of this project. The old event builder still have bugs (the bugs found in it has been fixed in this commit), and we won't maintain it anymore.

We use Gson deserializer system to make it works.

---------
Reviewed-by: xiaoACE <xiao_ace@icloud.com>
Tested-by: xiaoACE <xiao_ace@icloud.com>